gradio_GPT/requirements.txt
gradio>=4.0.0
Pillow>=9.0.0
openai>=1.0.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
numpy>=1.24.0
easyocr>=1.7.0
requests>=2.31.0
pandas>=2.0.0 